Update Person model and related dependencies
Updated the `Person` model entity and related files to support new attributes, username, dateOfBirth and gender. This also includes modifications to DTOs, Service and Controller classes, and the person repository to refactor usage from email to username. With these changes, we've added username as the unique identifier of a Person and introduced demographic fields. This is to align the Person model with new identification and demographic requirements. Also added a new Enum class for Gender to encapsulate possible gender representation. Furthermore, the import statements have been optimized across all affected classes. Credits were added to README.md for jwt implementation.