Add daily history and API tests
Added the functionality for persons to contain a list of their activities in the last seven days. This was done by modifying the history entity to handle daily activities and updating the PersonResponseDetailedDTO to return this information.
In addition, multiple API tests were added for users and authentication routes to ensure proper functionality of the endpoints.
Changes were also made to PersonController and DatabaseService classes to accommodate the newly added functionality. Storage location for uploaded files was moved to a local directory for easier access and management.