Add detailed history data with exercise records and improve password security.
This commit includes the implementation of more detailed history records in DTOs. It now provides History and Day objects with exercise records based on provided UUIDs. Also, make adjustments in the Day and History entity, transferring exerciseIDs from Day to History. Additionally, it calculates the strength of BCryptPasswordEncoder dynamically based on the number of CPU cores for better password security. Also, this commit adds the capability to find the history for a given person's ID. It creates relationships among Person, History, and Day and structures repositories for them. Moreover, it changes the UserDetailsService to use the UserDetails interface rather than directly accessing the UserDetails class and replaces the deprecated AbstractHttpConfigurer::disable with more proper alternatives. These changes improve the performance of the application and strength of password encoding adding functional richness to the history tracking of patients.

Overall, these adjustments provide a comprehensive, patient-specific history overview enhancing cognitive behavioral therapy possibilities in the Podotherapy Application.